Maintenance: use fine-grained lodash-imports
Lodash is a huge and chunky module. If we directly specify which methods we need from it, we can avoid a bit of the always-keep-it-up-to-date treadmill.

Further investigation could be done to see if we really even need the clone/isEqual methods from lodash, but at least with this PR change we put a soft fence around it to keep from panicking if lodash has security incidents.

Not a big deal if you don't want to drop it, since jest#core still needs it, but I thought I'd provide a PR