Feature: `importOrderMergeDuplicateImports` (#19)
This PR implements import declaration merging for duplicate imports.

I had to adjust getCodeFromAST so that we could still provide the original nodes, even if some of the nodes were being deleted.
It leverages the existing chunking logic, so if there's a side-effect import, or a prettier-ignore import, it won't merge across that boundary.
I did not implement conversion of import type { stuff } into import { type stuff } as it was not obvious when this would be wise. If we want to control that with an additional config option, something like importOrderMergeTypeStyle: "none" | "to-normal-import" | "unmerge" I could be convinced. I forsee people wanting opposite sides of the coin for that one.


* Add TypeScript enforcement to not-forgetting our Options schema!

* Change syntax of getCodeFromAst to allow for node deletion

* Github-linguist doesn't understand `ecmascript 6`

Just use javascript or js

* Feature: `importOrderMergeDuplicateImports` - Fixes #4

* switch from toMatchInlineSnapshot to toEqual

* PR Feedback: rename `regular` imports to `value` imports

* PR Feedback: Readme wording

* PR Feedback: rename getCodeFromAst parameters

Try to improve the naming and docs for the parameters, particularly the nodes vs importNodes distinciton

* PR Feedback: import-flavor @returns documentation

* PR Feedback: hoist hasIgnoreNextNode for reuse

* PR Feedback: Add strong types for ChunkType/FlavorType

* PR Feedback: deleteContext -> nodesToDelete

* PR Feedback: comment-update: once you mutate nodes, lines/locations are stale!

* Add a few tests and comments