[Breaking] Update dependencies, jest -> vitest, build for Node 16+ (#59)
I'm preparing to address
https://github.com/IanVS/prettier-plugin-sort-imports/issues/22, and as
the first step I've created a `next` branch for us to begin working on
which will become version 4.0.

This PR also lays some groundwork by updating some dependencies and
modernizing a few things:

- Typescript 4.8.4 -> 5.0.4
- Built files ES5 -> ES2021 (Node 16+)
- Jest -> Vitest (tests run in 5 seconds vs 14)
- `.npmignore` -> `files` in package.json (safer)
- Some other minor dep updates

Node 14 is going to be EOL at the end of this month, and I don't think
there's a reason to keep shipping ES5, which is very old by now, so I
set up the tsconfig to support Node 16. That's actually the only real
user-facing change in this PR, and I'm willing to reconsider it if
anyone feels differently about it.