[Breaking] Remove importOrderCaseInsensitive option (#64)
Ref #22 

This removes the `importOrderGroupNamespaceSpecifiers` option in an
effort to simplify the plugin and make it easier to understand and set
up. There's no replacement currently, but if we hear from lots of folks
who really want this, I think we could re-implement it using a special
term in `importOrder`. I haven't done that here, as I would prefer to
keep it simple.