Feature: drop javascript-natural-sort dep (#28)
This PR removes our reliance on natural-sort, an [npm package last
updated 9 years
ago!](https://www.npmjs.com/package/javascript-natural-sort)

Instead this uses
[`String.localeCompare(…)`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/localeCompare)
which is a built-in method supported on all Node versions we target.
Further this is a standard String api, so some shim implementation will
be present even if the user isn't running on Node! (eg. `deno` or other
runtimes)

Co-authored-by: Ian VanSchooten <ian.vanschooten@gmail.com>