Replace `importOrderCombineTypeAndValueImports` with `importOrderTypeScriptVersion` (#67)
Ref https://github.com/IanVS/prettier-plugin-sort-imports/issues/22

This removes the `importOrderCombineTypeAndValueImports` option, and
defaults merging to true for Flow projects (technically anything other
than TypeScript, but Flow is the only one).

For TypeScript, we need to be sure that we can use the `import {type
Foo} from …` syntax, which was only added in 4.5. To that end, this adds
a new option, `importOrderTypeScriptVersion`, which the user should set
to the version of TS they're using in their project, and we will use
that to generate appropriate import statements. And possibly in the
future we can make other adjustments based on the TS version as well.
Detecting the version automatically is tricky, but maybe we can try
doing that sometime in the future as well.

By default, we set the TS version to something ridiculously low,
`1.0.0`, meaning that by default we won't combine type and value unless
the user sets their version.