Clean-up lodash.clone dep, as it serves no purpose (#74)
I believe the usage of `lodash.clone` used to be more prevalent (and
necessary). But at this point we only had 1-remaining call-site, and I
traced it, and don't see any mutating methods along the way, so I'm
relatively confident this isn't necessary.

I also tested `JSON.parse(JSON.stringify(…))` and that passed the tests
too, so if we'd prefer to use that instead, I can do so, just let me
know!