recommend using ^react$ for sorting react to the top. (#126)
Hi there! 

I copy and pasted this from the readme and got confused when I found
`react-query` and other `react-` imports also getting sorted to the top.
I think it would be a simple change to recommend people use `^react$`
instead of just `react` for sorting react to the top.

Also, thanks so much for this plugin. It's incredible and exactly what
I've been looking for!