Regenerate lockfile (#129)
Lockfiles are great for keeping unexpected changes in dependencies from
breaking your app all of a sudden, but they also tend to cause
stagnation and duplicated packages (because packages that are added over
time use different versions of sub-dependencies). For that reason, it's
good to periodically trash the lockfile and re-generate it fresh, which
is what this PR does.

I had to tweak some code to make types happy, but our tests are still
passing so I think we're still all good.