Fix vue 2.7 support (#173)
Fixes https://github.com/IanVS/prettier-plugin-sort-imports/issues/162

Older versions (not sure which ones exactly) of vue/compiler-sfc do not
include a `loc` property on each block, but rather give the offsets
directly on the block itself. This accounts for those differences, and
causes the reproduction provided in the issue to succeed.