Add support for return in Astro component script (#176)
Addresses
https://github.com/IanVS/prettier-plugin-sort-imports/issues/96#issuecomment-2198005393

In Astro, it's possible to [return a
Response](https://docs.astro.build/en/guides/server-side-rendering/#return-a-response-object)
from inside of a component script, without wrapping it in a function.
This doesn't look like valid JavaScript, so the babel parser choked on
it.

This change checks to see if the file was parsed with the astro plugin,
and if so, enables the `allowReturnOutsideFunction` parser option.

This is admittedly a bit of a band-aid. I think the more correct
approach is to parse the code into AST using the configured parser,
rather than always using our version of babel. This is what some other
prettier plugins do. Some, like tailwindcss, goes so far as to load a
list of compatible other prettier plugins to enable cross-compatibility.

But, that's a much bigger change that will potentially require a
breaking change version bump, so I didn't tackle it here.

The other thing I did do here, is update the README a bit. I added
`Astro` to our supported frameworks as experimental (there may be other
parser options we need to set), and also included support for Svelte,
which we've always supported, as far as I know, but don't really test
for (it would be good to make some tests sometime).