Throw an error when no matching sortOrder group (#192)
Closes #190 

This adds a more understandable error message when we encounter an
import that does not fit into any group within `sortOrder`. I think the
only time that should happen is when the user adds groups with
`<TYPES>^.....` but no fallback `<TYPES>` bucket. We could try to shove
one in somewhere automatically like we do with `<THIRD_PARTY_MODULES>`,
but that could be pretty confusing and I don't know where it should go
in that case. This also includes a hint about how to fix the error.