Short-circuit before parsing (#224)
In attempting to troubleshoot
https://github.com/IanVS/prettier-plugin-sort-imports/issues/219, I
realized that disabling the plugin using overrides and `importOrder: []`
does not prevent parsing errors. Moving this short-circuit check up
before we attempt to parse will not only solve that problem but also
improve performance a little by avoiding parsing files that we're just
going to short-circuit anyway.