Be more tolerant of babel parse errors (#230)
Closes #219
Closes #222 

This change makes our babel parsing and node traversal more tolerant of
errors, matching the settings and behavior of Prettier itself. Now,
rather than this plugin causing an error and failing to format the file,
it will continue to attempt to sort imports. It may still fail on
invalid code, but in those cases prettier itself would as well, so we
shouldn't end up in a case where prettier runs but imports aren't
sorted, or where prettier would have run as long as this plugin wasn't
being used.

I chose not to emit a warning when invalid syntax is encountered, again
matching the behavior or prettier. I don't think it's the responsibility
of a import sorting plugin to alert the user to invalid syntax, since it
may be intentional and handled later-on in their build pipeline (for
instance, Astro files allow returning from the top-level).