Clean up options & remove explicit function types (#239)
This is a bit of a cleanup PR, and does two main things:

- Removes function types from `src/types.ts`, and defines the types
directly on the functions themselves. This was a holdover from before
the conversion of the project to typescript, when we had to have a
separate types file, but it's backwards from how types are normally
defined. It's clearer and more typesafe to define the types directly on
the functions, which this PR does now.
- Improves the options that we present to prettier (for it to show in
`prettier --help`, supposedly, though I'm not sure whether that actually
works or not).

Note: this doesn't have any impact on the types that we expose to users
(which is manually maintained in `types/index.d.ts`